+++ CHANGELOG.md
@@ -19,7 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
   Thanks to @babaruh and @Kataane for working on this improvement.
 
 ### Added
-- `Equals` method on `AbsolutePath` and `LocalPath` that accepts an alternate comparer (see `TruePath.Comparers.PlatformDefaultPathComparer` and `StrictStringPathComparer`). Thanks to @babaruh and @Kataane for working on this improvement.
+- `Equals` method on `AbsolutePath` and `LocalPath` that accepts an alternate comparer (see `PlatformDefaultComparer` and `StrictStringComparer` static comparers on both types). Thanks to @babaruh and @Kataane for working on this improvement.
 
 ## [1.5.0] - 2024-09-22
 ### Fixed